+++ components/icons-channel/talk-square/index.ts
@@ -1,0 +1,18 @@
+import { defineCustomElement } from '../../utils'
+import { createIconClass } from '../create-icon-class'
+import templateHTML from './template.html'
+import type { TSinchIconChannelElement, TSinchIconChannelReact } from '../types'
+
+defineCustomElement('sinch-icon-channel-talk-square', createIconClass(templateHTML))
+
+declare global {
+  namespace JSX {
+    interface IntrinsicElements {
+      'sinch-icon-channel-talk-square': TSinchIconChannelReact,
+    }
+  }
+
+  interface HTMLElementTagNameMap {
+    'sinch-icon-channel-talk-square': TSinchIconChannelElement,
+  }
+}