+++ components/illustration/create-illustration-class.ts
@@ -1,5 +1,7 @@
-import { getIntegerAttribute, NectaryElement, updateAttribute, updateIntegerAttribute } from '../utils'
+import { getIntegerAttribute, getLiteralAttribute, NectaryElement, updateAttribute, updateIntegerAttribute, updateLiteralAttribute } from '../utils'
 import illustrationStylesHtml from './illustration-styles.html'
+import { backgroundValues, valignValues } from './utils'
+import type { TSinchIllustrationBackground, TSinchIllustrationVAlign } from './types'
 
 const DEFAULT_SIZE = 256
 const MIN_SIZE = 16
@@ -8,7 +10,7 @@ const MAX_SIZE = 2048
 export const createIllustrationClass = (templateHTML: string): CustomElementConstructor => {
   const template = document.createElement('template')
 
-  template.innerHTML = illustrationStylesHtml + templateHTML
+  template.innerHTML = `${illustrationStylesHtml}<div id="wrapper">${templateHTML}</div>`
 
   return class extends NectaryElement {
     $svg: SVGElement
@@ -34,6 +36,22 @@ export const createIllustrationClass = (templateHTML: string): CustomElementCons
 
     get size() {
       return getIntegerAttribute(this, 'size', DEFAULT_SIZE)
+    }
+
+    get background(): TSinchIllustrationBackground | null {
+      return getLiteralAttribute(this, backgroundValues, 'background', null)
+    }
+
+    set background(value: TSinchIllustrationBackground | null) {
+      updateLiteralAttribute(this, backgroundValues, 'background', value)
+    }
+
+    get valign(): TSinchIllustrationVAlign | null {
+      return getLiteralAttribute(this, valignValues, 'valign', null)
+    }
+
+    set valign(value: TSinchIllustrationVAlign | null) {
+      updateLiteralAttribute(this, valignValues, 'valign', value)
     }
 
     connectedCallback() {