+++ tests/apps/angular/src/components/Input/Input.component.ts
@@ -1,6 +1,8 @@
 import { Component } from '@angular/core'
 import '@sinch-engage/nectary/input'
 import '@sinch-engage/nectary/help-tooltip'
+import '@sinch-engage/nectary/icon-button'
+import '@sinch-engage/nectary/icons/calendar-today'
 
 @Component({
   selector: 'input-component',
@@ -19,6 +21,7 @@ export class InputComponent {
   placeholderText: string | null
   tooltipText: string | null
   isDisabled: boolean
+  hasRightButton: boolean
 
   constructor() {
     const url = new URL(location.href)
@@ -32,6 +35,7 @@ export class InputComponent {
     this.placeholderText = url.searchParams.get('placeholder')
     this.tooltipText = url.searchParams.get('tooltip')
     this.isDisabled = url.searchParams.get('disabled') != null
+    this.hasRightButton = url.searchParams.get('right') != null
   }
 
   onChange(e: Event) {