+++ tests/apps/vue/src/components/Segment.vue
@@ -11,6 +11,10 @@
     <sinch-icon-button v-if="hasInfo" small slot="info">
       <sinch-icon-apps slot="icon"></sinch-icon-apps>
     </sinch-icon-button>
+    <div v-if="hasPreview" slot="preview" style="display: flex;flex-direction: column;background-color: #F1F3F4;align-items: center;justify-content: center;height: 100%;">
+      <span style="font-size: 18px">Replace me!</span>
+      <span style="font-size: 12px">Im a template component</span>
+    </div>
     <sinch-checkbox v-if="hasAction" text="Checkbox" slot="action"></sinch-checkbox>
     <sinch-button v-if="hasAction" text="Cancel" type="secondary" slot="action" small></sinch-button>
     <sinch-button v-if="hasAction" text="Ok" type="primary" slot="action" small></sinch-button>
@@ -50,6 +54,9 @@ export default {
     },
     hasInfo() {
       return this.search.get('info') !== null
+    },
+    hasPreview() {
+      return this.search.get('preview') !== null
     }
   },
   methods: {